+++ aiobungie/__init__.py
@@ -46,7 +46,6 @@ from aiobungie import url
 from aiobungie.client import Client
 from aiobungie.error import *
 from aiobungie.internal import iterators
-from aiobungie.internal.assets import Image
 from aiobungie.internal.enums import *
 from aiobungie.internal.iterators import *
 from aiobungie.rest import *