+++ tests/aiobungie/crates/test_character.py
@@ -31,7 +31,6 @@ import pytest
 
 import aiobungie
 from aiobungie import crates
-from aiobungie.internal import assets
 
 
 class TestDye:
@@ -156,8 +155,8 @@ class TestCharacter:
             light=1310,
             gender=aiobungie.Gender.MALE,
             race=aiobungie.Race.EXO,
-            emblem=assets.Image(path="emblempath.jpg"),
-            emblem_icon=assets.Image(path="emblemiconpath.jpg"),
+            emblem=aiobungie.builders.Image(path="emblempath.jpg"),
+            emblem_icon=aiobungie.builders.Image(path="emblemiconpath.jpg"),
             emblem_hash=998877,
             last_played=datetime.datetime(2021, 9, 1),
             total_played_time=2020,
@@ -190,13 +189,14 @@ class TestCharacter:
         assert model.title_hash is None
 
     def test_emblem(self, model: crates.Character) -> None:
-        assert model.emblem == assets.Image(path="emblempath.jpg")
+        assert model.emblem == aiobungie.builders.Image(path="emblempath.jpg")
 
     def test_emblem___str__(self, model: crates.Character) -> None:
-        assert str(model.emblem) == str(assets.Image(path="emblempath.jpg"))
+        assert str(model.emblem) == str(aiobungie.builders.Image(path="emblempath.jpg"))
 
     def test_emblem_icon(self, model: crates.Character) -> None:
         assert (
             model.emblem_icon is not None
-            and model.emblem_icon.url == assets.Image(path="emblemiconpath.jpg").url
+            and model.emblem_icon.create_url()
+            == aiobungie.builders.Image(path="emblemiconpath.jpg").create_url()
         )